Remove '-plugin' from infered plugin name
