+++ lib/core/backend/index.ts
@@ -1,3 +1,5 @@
 export * from './backend';
 
 export * from './backendCluster';
+
+export * from './backendPlugin';
